Switch bpaf to using derive macro
